Convert `FeedParser` into function
