+++ utils/__init__.py
@@ -1,0 +1,1 @@
+from feed_parsing import extract_jobs
